chore: add volume group discount fields to master entities and gold reward eligibility to master shops

+++ database/migrations/2026_05_15_090114_add_vol_gr_discount_in_master_shop_stats_table.php
@@ -1,5 +1,11 @@
 <?php
 
+/*
+ * Author: Raul Perusquia <raul@inikoo.com>
+ * Created: Fri, 15 May 2026 18:17:51 Malaysia Time, Kuala Lumpur, Malaysia
+ * Copyright (c) 2026, Raul A Perusquia Flores
+ */
+
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -9,14 +15,13 @@ return new class () extends Migration {
     {
         if (!Schema::hasColumn('master_shop_stats', 'number_master_families_with_vol_gr_discount')) {
             Schema::table('master_shop_stats', function (Blueprint $table) {
-                $table->unsignedInteger('number_master_families_with_vol_gr_discount')->default(0);
+                $table->unsignedSmallInteger('number_master_families_with_vol_gr_discount')->default(0);
             });
         }
     }
 
     public function down(): void
     {
-
         if (Schema::hasColumn('master_shop_stats', 'number_master_families_with_vol_gr_discount')) {
             Schema::table('master_shop_stats', function (Blueprint $table) {
                 $table->dropColumn('number_master_families_with_vol_gr_discount');
